requirements.txt
PyYAML>=6.0
python-dateutil>=2.9
pytest>=8.0
openbb>=4.0
python-dotenv>=1.0
